Move internal function resizeTty to container_internal
